Add makefile and before-commit.sh (#103)
* Add makefile and before-commit.sh

* Fix the catalog structure of cmd directory

* Update the .toml and .lock files

* Update the dockerfile with proper file structure of the cmd directory
